requirements.txt
streamlit
boto3==1.28.59
botocore==1.31.59
opensearch-py
camelot-py[cv]
